Preprocess rails assets imported via absolute path
If a processed file used an absolute path without an extension to import an
asset that would normally require preprocessing via the sassc-rails gem (such
as assets using ERB templates or `.css.s[a|c]ss` files, these files would fail
to process with a file not found error.

This change checks if the directory component of the import path is a valid
directory. If so, that directory is added to the list of paths to search for
files with extensions requiring preprocessing, and preprocesses those files
before handing them off to the native sassc library.