Preprocess rails assets imported with extensions.
If a processed file imported an erb asset with the erb extension in the import
argument, these files would not be preprocessed by the sassc-rails and the
native sassc functions would fail to compile the assets.

This change checks if the import argument ends with any of the extensions known
by the sassc preprocessor, and if it does, applies the appropriate
preprocessing to that file.